Refactor: Update NFCTagManager and related code
