+++ backend/ntags/views.py
@@ -2,7 +2,7 @@ from django.contrib import messages
 from django.shortcuts import redirect
 from django.utils.translation import gettext_lazy as _
 
-from .utils import get_nfc_tag_model
+from . import get_nfc_tag_model
 
 def link(request):
     """
@@ -18,10 +18,7 @@ def link(request):
         ntag = NFCTag.objects.get_from_mirror(mirrored_values)
         return redirect(ntag.linked_item.url)
 
-    except ValueError as e:
+    except Exception as e:
         messages.error(request, str(e))
         return redirect('/')
 
-    except NFCTag.DoesNotExist as e:
-        messages.error(request, str(e))
-        return redirect('/')
